refactor: second pass — extract data module, break god methods, unify patterns
match_engine.py (1279→~950 lines):
- Extract 330 lines of static data (MOVES, SIGNATURE_MOVE_POOLS, ARCHETYPE_FINISHERS,
  spot pools, crowd reactions, etc.) into new core_engine/match_data.py
- Re-export all symbols for backward compatibility
- Fix duplicate near-fall bug (was generating 2 near-falls per tick)
- Remove dead code (_should_switch_control pass statement in multi-person)

world_ticker.py (1217→~1210 lines):
- Break 88-line _wrestler_lifecycle() into 4 focused dispatchers:
  _lifecycle_annual(), _lifecycle_seasonal(), _lifecycle_weekly(),
  _lifecycle_daily_ring_rust()
- Each handles one cadence (annual/seasonal/weekly/daily) with clear boundaries

show_service.py:
- Unify duplicated promo segment booking (was 22 lines with identical
  book_promo_segment() calls in both if/else branches → 10 lines)
- Extract FINISH_TYPES and FINISH_TYPE_WEIGHTS constants

promo_service.py:
- Build ARCHETYPE_TEMPLATES consolidated dict from OPENERS/BODIES/CLOSERS
  (single source of truth per archetype)
- Extract NEGATIVE/POSITIVE_LIFE_EVENT_TYPES to module-level sets
- Extract 6 more constants: SEVERE_EVENT_THRESHOLD, CATASTROPHIC_EVENT_THRESHOLD,
  GRIEF_MORALE_IMPACT, MIN_LLM_RESULT_LENGTH, MIN/MAX_GOOD_PROMO_WORDS

match_aftermath.py:
- Unify _record_history() winner/loser blocks (was 2 identical loops → 1 parameterized)
- Use _get_relationship() helper in _update_relationships() (was duplicating the query)
- Use _get_wrestler() in pre-fetch loop
- Move CHEMISTRY_*/DEFAULT_MATCH_RATING/MATCH_RATING_SCALE to module-level constants

All 293 tests pass with no regressions.
